Switch from patching Launchpad to forking
We found out that we may not rely on patch-package in a library,
as it is intended to work exclusively in top level projects and
not recommended to be used in a library. References:

- https://github.com/ds300/patch-package/issues/198
- https://github.com/ds300/patch-package/issues/26

For that reason we had to create a new repository containing the
patched version of the Launchpad package's latest build. We
patched the build instead of forking the whole package and
patching the source for practical reasons, please see the package
README if you're curious about those reasons.