Implement option to create wallet from ctor args
Add an option to initialize a wallet by supplying the constructor
arguments of Secp256k1HdWallet and IxoAgentWallet classes.

Before this, wallet initialization was only possible by supplying
either a mnemonic or a serialized wallet, which are slow as they
both cause the generation of wallets from scratch using the
mnemonic, while the initialization via constructor arguments will
happen instantly.