Reimplement wallet serialization
The old serialization method relied on the serialization features
on underlying Secp256k1HdWallet instances, which is suboptimal as
deserialization involved regenerating the wallet from scratch
using the mmemonic, which is time consuming and unnecessary. It
also is completely custom and isn't designed to work with native
JSON.stringify and JSON.parse serialization methods.

This implementation addresses both problems. Thanks to it:

- Wallets can now be deserialized instantly

- Serialization now works via native JSON.stringify and JSON.parse

- It is now possible to obtain a representation of the wallet's
  internal state via it's "toJSON" method (required by the
  JSON.stringify integration.) This can be useful whenever one
  needs to access / embed the wallet's state without dealing with
  a class instance. (Example: Embedding wallet state into a global
  app state which doesn't support class instances)