Swap "entity" with "project", "template" & "cell"
A lot of entities such as projects, templates and cells are being
stored as abstract "entity" records in ixo backends. But that is
in fact too abstract for this SDK. SDK users shouldn't worry about
our abstractions too much and they should work with more concrete
types instead.

For that reason we are removing all "entity" methods (or making
them undocumented) and introducing the more concrete "project",
"template" and "cell" methods instead.