Implement the "dashifyUrls" client option
Some clients do not support URLs with underscores in them, most
significantly the HTTP client of the Android SDK version required
by React Native. Since at the moment there are a lot of entities
with underscored URLs hardcoded in the ixo blockchain, we felt the
need to provide this option to automatically dashify URLs.