Revert "Improve wallet serialization"
This reverts commit 053901478d2cbc8968dfc9134e8d008b5405fad6.

I've changed my mind. The plain object version of the wallet is
provided mostly for serialization purposes. The serialization
output is supposed to be as small as possible, so it is arguably
better to have it contain the least amount of data that is
sufficient for a proper deserialization.

According to this logic, we don't need the did information in he
plain object version of the wallet as it's redundant.