Make the wallet DID prefix configurable
Wallets can be generated by other services than this client SDK
and they may use different DID prefixes than the default prefix of
"did:ixo:" used here.

While importing a wallet using a mnemonic, only the body part of
the DID can be reconstructed using the mnemonic and not the prefix
part. So we need to accept the prefix as an additional parameter
to the mnemonic in order to fully support wallets generated by
other services.